Split consistent-trees tests.
